requirements.txt
python-docx>=0.8.11
markdown>=3.4.1
PyMuPDF>=1.21.0
openai>=1.0.0
langdetect>=1.0.9
streamlit>=1.28.0
streamlit-cookies-manager>=0.2.0
cryptography>=41.0.0
requests>=2.28.0
tiktoken>=0.7.0
google-genai